Renamed uses of resultRow to datasetRow
