--- assets/js/vue/components/search/ResultRow.vue
@@ -1,90 +1,0 @@
-<template>
-    <b-card class="card-product">
-        <b-link class="search-result"
-                :href="url" target="_blank">
-            <div>
-                <span class="badge badge-secondary" v-if="resultRowData.availabilityStatus === 0">Identified</span>
-                <span class="badge badge-primary" v-else-if="resultRowData.availabilityStatus === 2 || resultRowData.availabilityStatus === 4">Submitted</span>
-                <span class="badge badge-danger" v-else-if="resultRowData.availabilityStatus === 5 || resultRowData.availabilityStatus === 8">Restricted</span>
-                <span class="badge badge-success" v-else-if="resultRowData.availabilityStatus === 7 || resultRowData.availabilityStatus === 10">Available</span>
-            </div>
-            <b-card-title>{{ resultRowData.title }}</b-card-title>
-            <b-card-text class="d-flex justify-content-between" >
-                <div v-if="Object.keys(resultRowData.datasetSubmission).length > 0">
-                    <div v-if="resultRowData.datasetSubmission.authors">
-                        Authors: {{ resultRowData.datasetSubmission.authors }}
-                    </div>
-                    <div v-if="resultRowData.acceptedDate">
-                        Published on {{ resultRowData.acceptedDate }}
-                    </div>
-                    <div v-if="resultRowData.fileFormat">
-                        File Format: {{ resultRowData.fileFormat }}
-                    </div>
-                </div>
-                <div>
-                    <div v-if="resultRowData.doi.doi">
-                        DOI: {{ resultRowData.doi.doi }}
-                    </div>
-                    <div>
-                        UDI: {{ resultRowData.udi }}
-                    </div>
-                    <div v-if="resultRowData.fileSize">
-                        File Size: {{ resultRowData.fileSize }}
-                    </div>
-                </div>
-            </b-card-text>
-        </b-link>
-    </b-card>
-</template>
-
-<script>
-    export default {
-        name: "ResultRow",
-        props: {
-            resultRowData: {
-                type: Object
-            }
-        },
-        data: function () {
-            return {
-                url: Routing.generate("pelagos_app_ui_dataland_default", {'udi' : this.resultRowData.udi } )
-            }
-        }
-    }
-</script>
-
-<style scoped>
-    div.bootstrap p {
-        margin-bottom: 0 !important;
-    }
-
-    div.bootstrap h4 {
-        font-size: 1.3rem !important;
-    }
-
-    .card-product {
-        margin-bottom: 1rem;
-    }
-
-    .card-product:hover .btn-overlay {
-        opacity: 1;
-    }
-
-    .card-product:hover {
-        -webkit-box-shadow: 0 4px 15px rgba(153, 153, 153, 0.3);
-        box-shadow: 0 4px 15px rgba(153, 153, 153, 0.3);
-        -webkit-transition: .5s;
-        transition: .5s;
-        cursor: pointer;
-    }
-
-    a.search-result {
-        font-size: 1.05em;
-        text-decoration: none !important;
-        color: black !important;
-    }
-
-    .search-result a:hover {
-        text-decoration: none !important;
-    }
-</style>
